seeder/requirements.txt
psycopg2-binary==2.9.10
neo4j==5.25.0
